# Conflicts:
#	connector/src/main/java/org/geysermc/connector/network/translators/world/block/BlockTranslator.java

+++ connector/src/main/java/org/geysermc/connector/network/translators/java/world/JavaTradeListTranslator.java
@@ -140,7 +140,7 @@ public class JavaTradeListTranslator extends PacketTranslator<ServerTradeListPac
 
         NbtMapBuilder builder = NbtMap.builder();
         builder.putByte("Count", (byte) (Math.max(itemData.getCount() + specialPrice, 1)));
-        builder.putShort("Damage", itemData.getDamage());
+        builder.putShort("Damage", (short) itemData.getDamage());
         builder.putString("Name", itemEntry.getBedrockIdentifier());
         if (itemData.getTag() != null) {
             NbtMap tag = itemData.getTag().toBuilder().build();
